restructure: move GameObjects to `game-objects` folder
