+++ src/game-objects/BomberMan/assets/images/back/Bman_B_f06.png
