+++ src/game-objects/BomberMan/assets/images/front/Bman_F_f02.png
