+++ src/game-objects/BomberMan/assets/images/left/Bman_L_f03.png
