+++ src/game-objects/BomberMan/assets/images/right/Bman_R_f02.png
