+++ src/game-objects/BomberMan/assets/images/right/Bman_R_f06.png
